ECS redesign. "Entities" are now just unsigned ints.
Each Component has a System with an AssocMap (think having benefits of
key->value like a map, but contiguous memory for fast iteration like a vector).
Each System contains all the components for that type.
A bit more work needs to be done (including a decision of what owns the
Systems (prolly a World)). After that, it is done!